0208 // Interceptor CommonInterceptor.java
// Request Logging
LogRequest.java
LogRequestRepository.java

// JPA Repositories
PublicChannelRepository.java
UserChannelRepository.java

// Using DB Data
Channel.tsx
ChannelHaed.tsx
ChannelItem.tsx

// Redux
channelRedux.ts

+++ src/main/java/io/iconic/backend/repository/PublicChannelRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PublicChannelRepository extends JpaRepository<PublicChannel, Long> {
+public interface PublicChannelRepository extends JpaRepository<PublicChannel, Integer> {
 
     Optional<PublicChannel> findPublicChannelBypChanIdx(int pChanIdx);
